move related apps field into components folder, add data entry instructions message
